requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pypdf
pillow
pytesseract
pdf2image
langchain
chromadb
pytest
boto3
